test: relocate test
