+++ test/unit/modules/plugin/helper/plugin.js
@@ -1,0 +1,12 @@
+
+import path from 'path';
+import config from '../../../partials/config';
+
+const authProfileConf = {
+  ...config(),
+  auth: {
+    [`${path.join(__dirname, '../../../partials/plugin/authenticate')}`]: { }
+  }
+};
+
+export default authProfileConf;
