--- test/unit/partials/config/plugin.js
@@ -1,10 +1,0 @@
-
-import path from 'path';
-import config from './index';
-
-module.exports = {
-  ...config(),
-  auth: {
-    [`${path.join(__dirname, '../plugin/authenticate')}`]: { }
-  }
-};
